refactor: move axelard commands
In preparation to migrating to cosmos-sdk >=v0.42.4 this change adopts a part of the new app structure to keep breaking changes as small as possible
