+++ cmd/axelard/cmd/gencontracts.go
@@ -1,4 +1,4 @@
-package main
+package cmd
 
 import (
 	"fmt"
@@ -81,7 +81,7 @@ func SetGenesisEthContractsCmd(
 	cmd.Flags().StringVar(&burnableFile, "burnable", "", "Path to the burner contract ABI.")
 
 	cmd.Flags().String(cli.HomeFlag, defaultNodeHome, "node's home directory")
-	cmd.Flags().String(CliHomeFlag, defaultClientHome, "client's home directory")
+	cmd.Flags().String(cliHomeFlag, defaultClientHome, "client's home directory")
 
 	return cmd
 }
