+++ cmd/axelard/cmd/gensnapshot.go
@@ -1,10 +1,10 @@
-package main
+package cmd
 
 import (
 	"fmt"
 	"time"
 
-	stakingTypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+	snapshotTypes "github.com/axelarnetwork/axelar-core/x/snapshot/types"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/server"
@@ -14,16 +14,15 @@ import (
 	"github.com/tendermint/tendermint/libs/cli"
 )
 
-// SetGenesisStakingCmd returns set-genesis-chain-params cobra Command.
-func SetGenesisStakingCmd(
+// SetGenesisSnapshotCmd returns set-genesis-chain-params cobra Command.
+func SetGenesisSnapshotCmd(
 	ctx *server.Context, cdc *codec.Codec, defaultNodeHome, defaultClientHome string,
 ) *cobra.Command {
-	var unbond string
-	var max uint16
+	var period string
 
 	cmd := &cobra.Command{
-		Use:   "set-genesis-staking",
-		Short: "Set the genesis parameters for the staking module",
+		Use:   "set-genesis-snapshot",
+		Short: "Set the genesis parameters for the snapshot module",
 		Args:  cobra.ExactArgs(0),
 		RunE: func(_ *cobra.Command, args []string) error {
 
@@ -35,30 +34,22 @@ func SetGenesisStakingCmd(
 			if err != nil {
 				return fmt.Errorf("failed to unmarshal genesis state: %w", err)
 			}
+			genesisSnapshot := snapshotTypes.GetGenesisStateFromAppState(cdc, appState)
 
-			var genesisStaking stakingTypes.GenesisState
-			if appState[stakingTypes.ModuleName] != nil {
-				cdc.MustUnmarshalJSON(appState[stakingTypes.ModuleName], &genesisStaking)
-			}
-
-			if unbond != "" {
-				period, err := time.ParseDuration(unbond)
+			if period != "" {
+				period, err := time.ParseDuration(period)
 				if err != nil {
 					return err
 				}
-				genesisStaking.Params.UnbondingTime = period
+				genesisSnapshot.Params.LockingPeriod = period
 			}
 
-			if max > 0 {
-				genesisStaking.Params.MaxValidators = max
-			}
-
-			genesisSnapshotBz, err := cdc.MarshalJSON(genesisStaking)
+			genesisSnapshotBz, err := cdc.MarshalJSON(genesisSnapshot)
 			if err != nil {
 				return fmt.Errorf("failed to marshal snapshot genesis state: %w", err)
 			}
 
-			appState[stakingTypes.ModuleName] = genesisSnapshotBz
+			appState[snapshotTypes.ModuleName] = genesisSnapshotBz
 
 			appStateJSON, err := cdc.MarshalJSON(appState)
 			if err != nil {
@@ -70,11 +61,10 @@ func SetGenesisStakingCmd(
 		},
 	}
 
-	cmd.Flags().StringVar(&unbond, "unbonding-period", "", "Time duration of unbonding (e.g., \"6h\").")
-	cmd.Flags().Uint16Var(&max, "max-validators", 0, "A positive integer representing the maximum number of validators (max uint16 = 65535)")
+	cmd.Flags().StringVar(&period, "locking-period", "", "Locking period for the snapshot module (e.g., \"6h\").")
 
 	cmd.Flags().String(cli.HomeFlag, defaultNodeHome, "node's home directory")
-	cmd.Flags().String(CliHomeFlag, defaultClientHome, "client's home directory")
+	cmd.Flags().String(cliHomeFlag, defaultClientHome, "client's home directory")
 
 	return cmd
 }
